Recreated plots to follow best practices in information graphics.
Base on what I read in Dona Wong's book on infographics, and also the
Harris book "Information Graphics", it seems that having grid markings
for horizontal bar plots is uncommon.

Also, following the approach used by Frakes and Fox in their 1995
paper, I put information about the number of responses (and sometimes
the mean and standard deviations) directly on the plots.